DS-260: Move DropdownMenu mock data to separate file
