+++ ngx-fudis/projects/ngx-fudis/src/lib/components/form/select/common/mock_data.ts
@@ -1,13 +1,12 @@
+/**
+ * This file contains mock data related to Select and Multiselect Components
+ */
+
 import { FudisSelectOption } from '../../../../types/forms';
 
 interface MockSelectOptionGroupData {
   country: string;
   options: FudisSelectOption<TestAnimalScience>[];
-}
-
-interface MockDropdownMenuGroupData {
-  country: string;
-  items: TestMenuItemActions[];
 }
 
 export type TestAnimalSound = {
@@ -30,19 +29,6 @@ export const defaultOptions: FudisSelectOption<TestAnimalSound>[] = [
   { value: 'value-6-gecko', label: 'Southern Titiwangsa Bent-Toed Gecko', sound: 'Gec-koooo!' },
 ];
 
-export const defaultMenuItems: TestMenuItemActions[] = [
-  {
-    label: 'First item',
-    disabled: false,
-  },
-  { label: 'Second item', disabled: true },
-  {
-    label: 'Third item with really long label to push it to the limit!',
-    disabled: false,
-  },
-  { label: 'Fourth item', disabled: false },
-];
-
 export const multiselectChipListMockData: FudisSelectOption<object>[] = [
   { value: 'hereford', label: 'Hereford' },
   { value: 'texas-longhorn', label: 'Texas Longhorn' },
@@ -54,11 +40,6 @@ export type TestAnimalScience = {
   value: string;
   label: string;
   subLabel: string;
-  disabled?: boolean;
-};
-
-export type TestMenuItemActions = {
-  label: string;
   disabled?: boolean;
 };
 
@@ -970,46 +951,6 @@ export const smallGroupedMockData: MockSelectOptionGroupData[] = [
         value: 'e2fa6f0a-632a-485f-8ccf-b984311fe3b4',
         label: 'Mountain lion',
         subLabel: 'Felis concolor',
-      },
-    ],
-  },
-];
-
-export const smallDropdownMenuGroupedMockData: MockDropdownMenuGroupData[] = [
-  {
-    country: 'Netherlands',
-    items: [
-      {
-        label: 'Golden jackal',
-        disabled: false,
-      },
-      {
-        label: 'Mountain lion',
-        disabled: false,
-      },
-    ],
-  },
-  {
-    country: 'Brazil',
-    items: [
-      {
-        label: 'Small Indian mongoose',
-        disabled: true,
-      },
-      {
-        label: 'Falcon, prairie',
-        disabled: false,
-      },
-    ],
-  },
-  {
-    country: 'China',
-    items: [
-      {
-        label: 'Salmon pink bird eater tarantula',
-      },
-      {
-        label: 'Crane, sandhill',
       },
     ],
   },
